Move `_find_server_array` into the RightScaleBaseActor.
This method will be used by other RightScale actors to find Server
Arrays and act on them.
