+++ kingpin/actors/rightscale/base.py
@@ -32,9 +32,13 @@ below for using each actor.
   (defaults to https://my.rightscale.com)
 """
 
+from random import randint
 import collections
 import logging
 import os
+
+from tornado import gen
+import mock
 
 from kingpin.actors import base
 from kingpin.actors import exceptions
@@ -47,6 +51,16 @@ __author__ = 'Matt Wise <matt@nextdoor.com>'
 
 TOKEN = os.getenv('RIGHTSCALE_TOKEN', None)
 ENDPOINT = os.getenv('RIGHTSCALE_ENDPOINT', 'https://my.rightscale.com')
+
+
+class ArrayNotFound(exceptions.RecoverableActorFailure):
+
+    """Raised when a ServerArray could not be found."""
+
+
+class ArrayAlreadyExists(exceptions.RecoverableActorFailure):
+
+    """Raised when a ServerArray already exists by a given name."""
 
 
 class RightScaleBaseActor(base.BaseActor):
@@ -62,6 +76,72 @@ class RightScaleBaseActor(base.BaseActor):
                 'Missing the "RIGHTSCALE_TOKEN" environment variable.')
 
         self._client = api.RightScale(token=TOKEN, endpoint=ENDPOINT)
+
+    @gen.coroutine
+    def _find_server_arrays(self, array_name,
+                            raise_on='notfound',
+                            allow_mock=True,
+                            exact=True):
+        """Find a ServerArray by name and return it.
+
+        Args:
+            array_name: String name of the ServerArray to find.
+            raise_on: Either None, 'notfound' or 'found'
+            allow_mock: Boolean whether or not to allow a Mock object to be
+                        returned instead.
+            exact: Boolean whether or not to allow multiple arrays to be
+                   returned.
+
+        Raises:
+            gen.Return(<rightscale.Resource of Server Array>)
+            ArrayNotFound()
+            ArrayAlreadyExists()
+        """
+        if raise_on == 'notfound':
+            msg = 'Verifying that array "%s" exists' % array_name
+        elif raise_on == 'found':
+            msg = 'Verifying that array "%s" does not exist' % array_name
+        elif not raise_on:
+            msg = 'Searching for array named "%s"' % array_name
+        else:
+            raise exceptions.UnrecoverableActorFailure(
+                'Invalid "raise_on" setting in actor code.')
+
+        self.log.debug(msg)
+        array = yield self._client.find_server_arrays(array_name, exact=exact)
+
+        if not array and self._dry and allow_mock:
+            # Create a fake ServerArray object thats mocked up to help with
+            # execution of the rest of the code.
+            self.log.info('Array "%s" not found -- creating a mock.' %
+                          array_name)
+            array = mock.MagicMock(name=array_name)
+            # Give the mock a real identity and give it valid elasticity
+            # parameters so the Launch() actor can behave properly.
+            array.soul = {
+                # Used elsewhere to know whether we're working on a mock
+                'fake': True,
+
+                # Fake out common server array object properties
+                'name': '<mocked array %s>' % array_name,
+                'elasticity_params': {'bounds': {'min_count': 4}}
+            }
+            array.self.path = '/fake/array/%s' % randint(10000, 20000)
+            array.self.show.return_value = array
+
+        if array and raise_on == 'found':
+            raise ArrayAlreadyExists('Array "%s" already exists!' % array_name)
+
+        if not array and raise_on == 'notfound':
+            raise ArrayNotFound('Array "%s" not found!' % array_name)
+
+        # Quick note. If many arrays were returned, lets make sure we throw a
+        # note to the user so they know whats going on.
+        if isinstance(array, list):
+            for a in array:
+                self.log.info('Matching array found: %s' % a.soul['name'])
+
+        raise gen.Return(array)
 
     def _generate_rightscale_params(self, prefix, params):
         """Utility function for creating RightScale-style parameters.
